refactor: move common server/client types into separate crate
