+++ crates/node_types/prover/Cargo.toml
@@ -23,6 +23,7 @@ tokio = { workspace = true }
 log = { workspace = true }
 anyhow = { workspace = true }
 jmt = { workspace = true }
+prism-api = { workspace = true }
 prism-common = { workspace = true, features = ["test_utils"] }
 prism-storage = { workspace = true }
 prism-tree = { workspace = true }
