+++ crates/tree/Cargo.toml
@@ -12,6 +12,7 @@ readme.workspace = true
 
 [dependencies]
 # prism
+prism-api.workspace = true
 prism-common = { workspace = true, features = ["test_utils"] }
 prism-errors.workspace = true
 prism-keys.workspace = true
@@ -23,9 +24,6 @@ sha2.workspace = true
 # serde
 prism-serde.workspace = true
 serde.workspace = true
-
-# OAS spec
-utoipa.workspace = true
 
 # misc
 anyhow.workspace = true
